src/requirements.txt
boto3
kedro
pyyaml
python-dotenv